+++ Cargo.toml
@@ -305,6 +305,7 @@ pallet-asset-gateway = { path = "modules/ismp/pallets/asset-gateway", default-fe
 pallet-token-governor = { path = "modules/ismp/pallets/token-governor", default-features = false }
 pallet-state-coprocessor = { path = "modules/ismp/pallets/state-coprocessor", default-features = false }
 pallet-mmr = { path = "modules/trees/mmr/pallet", default-features = false }
+pallet-token-gateway = { path = "modules/ismp/pallets/token-gateway", default-features = false }
 
 # merkle trees
 pallet-mmr-runtime-api = { path = "modules/trees/mmr/pallet/runtime-api", default-features = false }
